new solution

+++ CPlusPlus-Basics/First Steps In Coding - Lab/09_Yard_Greening.cpp
@@ -1,13 +1,14 @@
 #include <iostream>
 
-int main() {
-    float a;
-    
+int main()
+{
+    double a;
+
     std::cin >> a;
 
-    float result = a * 7.61;
-    float off = 0.18 * result;
-    float total = result - off;
+    double result = a * 7.61;
+    double off = 0.18 * result;
+    double total = result - off;
 
     std::cout << "The final price is: " << total << " lv." << std::endl;
     std::cout << "The discount is: " << off << " lv." << std::endl;
